Adjusted Profile to try and reflect browse
Adjusted Profile to try and reflect elements of the browse game.
However, there are errors that are happening and it is not sorted
according to the individual's joined games. I'm not sure exactly how
good this code is and I wasn't able to debug it properly. However, I
tried to copy as many elements of the browse game as I could into the
profile.